feat(scripts): support no-entrypoint modules
This is a big one:

Prior to this commit, we didn't have a way to apply config based plugins to a
large portion of the codebase because those scripts didn't have a local `init`
or `new` method from which to trigger the loader.

I've defined a new mechanism to allow for loading plugins in these modules:

```
-- Set a global table so that we can detect modules that are loaded and need
-- plugin patching but which do not have an init method for hooks.
LPL_Additional_Paths = LPL_Additional_Paths or {}
LPL_Additional_Paths["/scripts/actions/animator_plugins.config"] = true
```

When the PluginLoader runs, if there are any unloaded plugins in the above map
it will load them.

Tested and working on actions/animator.lua and actions/npc.lua